unify the identical copies of TimerParams
