+++ hal/src/common/mod.rs
@@ -5,6 +5,7 @@ pub mod sercom;
 pub mod sleeping_delay;
 pub mod spi_common;
 pub mod time;
+pub mod timer_params;
 pub mod timer_traits;
 
 #[cfg(any(feature = "samd11", feature = "samd21"))]
