+++ hal/src/common/thumbv7em/timer.rs
@@ -3,6 +3,7 @@ use crate::hal::timer::{CountDown, Periodic};
 use crate::target_device::tc0::COUNT16;
 #[allow(unused)]
 use crate::target_device::{MCLK, TC2, TC3};
+use crate::timer_params::TimerParams;
 // Only the G variants are missing these timers
 #[cfg(feature = "min-samd51j")]
 use crate::target_device::{TC4, TC5};
@@ -168,59 +169,6 @@ impl TimerCounter<$TC>
     }
 }
         )+
-    }
-}
-
-/// Helper type for computing cycles and divider given frequency
-#[derive(Debug, Clone, Copy)]
-pub struct TimerParams {
-    pub divider: u16,
-    pub cycles: u32,
-}
-
-impl TimerParams {
-    pub fn new<T>(timeout: T, src_freq: u32) -> Self
-    where
-        T: Into<Hertz>,
-    {
-        let timeout = timeout.into();
-        let ticks: u32 = src_freq / timeout.0.max(1);
-        Self::new_from_ticks(ticks)
-    }
-
-    pub fn new_us<T>(timeout: T, src_freq: u32) -> Self
-    where
-        T: Into<Nanoseconds>,
-    {
-        let timeout = timeout.into();
-        let ticks: u32 = (timeout.0 as u64 * src_freq as u64 / 1_000_000_000_u64) as u32;
-        Self::new_from_ticks(ticks)
-    }
-
-    fn new_from_ticks(ticks: u32) -> Self {
-        let divider = ((ticks >> 16) + 1).next_power_of_two();
-        let divider = match divider {
-            1 | 2 | 4 | 8 | 16 | 64 | 256 | 1024 => divider,
-            // There are a couple of gaps, so we round up to the next largest
-            // divider; we'll need to count twice as many but it will work.
-            32 => 64,
-            128 => 256,
-            512 => 1024,
-            // Catch all case; this is lame.  Would be great to detect this
-            // and fail at compile time.
-            _ => 1024,
-        };
-
-        let cycles: u32 = ticks / divider as u32;
-
-        if cycles > u16::max_value() as u32 {
-            panic!("cycles {} is out of range for a 16 bit counter", cycles);
-        }
-
-        TimerParams {
-            divider: divider as u16,
-            cycles,
-        }
     }
 }
